refactor(param-id): extract the CasADi backend into param_id/casadi_backend.py
Pure move. paramID.py 3382 -> 3206 lines.

The CasADi symbolic machinery (functions_cache_key, build_functions, get_cost,
get_jac_cost, get_obs, plus require_casadi / as_casadi_column / content_hash)
now lives in its own module. Each function takes the param-id object as its
first argument and reads/writes state directly on it, so cost_symb, cost_func,
obs_meta and the rest stay instance attributes exactly as before.

casadi_backend imports nothing from paramID -- the dependency runs one way and
the module imports standalone (asserted in the verification below). This is the
same shape as param_id.plot_outputs, which is already called with a param-id
handle rather than importing the class.

OpencorParamID keeps thin delegating methods, so plot_outputs.py's calls to
get_cost_ca/get_obs_ca, the test-suite stubs implementing the same duck-typed
contract, and every direct test call are untouched.

require_casadi and as_casadi_column are re-bound to their previous private names
in paramID because the *generic* cost-assembly layer that stays there
(cost_calc, get_obs_output_dict) still builds SX directly and calls them.
content_hash had exactly one consumer, which moved with it, so it is gone from
paramID along with the now-unused hashlib import.

Verified: `./run_pytest.sh -k "casadi or series_interpolation or
sp_minimize_streams or multi_start or fsa"` gives byte-identical outcomes before
and after (39 tests, all passing, diffed by node id). Plus a standalone import
check that no paramID module is pulled in.
